Cleaned up IMU
