+++ urc_imu/package.xml
@@ -4,19 +4,13 @@
     <name>imu_driver</name>
     <version>2.0.1</version>
     <description>
-        Package to parse IMU strings and publish a very simple IMU message. Does not
-        require or use the GPSD deamon.
+        Package to parse IMU strings and publish a very simple IMU message.
     </description>
-
     <maintainer email="evenator@gmail.com">Ed Venator</maintainer>
-
     <license>BSD</license>
-
     <url type="website">http://ros.org/wiki/nmea_navsat_driver</url>
-
     <author email="eric@ericperko.com">Eric Perko</author>
     <author>Steven Martin</author>
-
     <exec_depend>geometry_msgs</exec_depend>
     <exec_depend>nmea_msgs</exec_depend>
     <exec_depend>rclpy</exec_depend>
